organize NavBar
