+++ packages/app-shell/src/NavBar/components/NavBarProducts/NavBarProducts.jsx
