+++ packages/app-shell/src/NavBar/index.jsx
@@ -26,9 +26,9 @@ import { fontFamily, fontWeightMedium } from '@bufferapp/ui/style/fonts';
 import Link from '@bufferapp/ui/Link';
 import DropdownMenu from '@bufferapp/ui/DropdownMenu';
 
-import BufferLogo from './BufferLogo';
-import NavBarMenu from './NavBarMenu/NavBarMenu';
-import NavBarProducts from './NavBarProducts/NavBarProducts';
+import BufferLogo from './components/BufferLogo';
+import NavBarMenu from './components/NavBarMenu/NavBarMenu';
+import NavBarProducts from './components/NavBarProducts/NavBarProducts';
 import { useUser } from '../context/User';
 import useOrgSwitcher from '../hooks/useOrgSwitcher';
